requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
attrs==25.3.0
frozenlist==1.7.0
idna==3.10
multidict==6.6.3
paho-mqtt==2.1.0
propcache==0.3.2
telepot==12.7
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
